+++ src/main/java/com/empresa/fichajes/service/security/AuthService.java
@@ -1,7 +1,7 @@
 package com.empresa.fichajes.service.security;
 
 import com.empresa.fichajes.entity.Usuario;
-import com.empresa.fichajes.Dao.UsuarioDao;
+import com.empresa.fichajes.dao.UsuarioDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
