+++ src/main/java/com/empresa/fichajes/service/security/JwtFilter.java
@@ -1,7 +1,7 @@
 package com.empresa.fichajes.service.security;
 
 import com.empresa.fichajes.entity.Usuario;
-import com.empresa.fichajes.Dao.UsuarioDao;
+import com.empresa.fichajes.dao.UsuarioDao;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
